add contant expand view item code
add contant expand view item code

+++ app/src/main/java/com/findg/component/ContactItemOperationListener.java
@@ -2,5 +2,5 @@ package com.findg.component;
 
 public interface ContactItemOperationListener {
 
-    void onAddUserClicked(int userId);
+    void onAddUserClicked(long userId);
 }
